pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parcl-crawler"
version = "0.1.0"
description = "Parcel-anchored real estate data crawler for Austin TX"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "click>=8.1",
    "duckdb>=1.0",
    "psycopg2-binary>=2.9",
    "pyarrow>=15.0",
    "pandas>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "black>=24.0",
    "ruff>=0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["parcl"]

[project.scripts]
parcl = "parcl.cli:main"
